LayoutTests/inspector/console/console-dir-expected.txt

CONSOLE MESSAGE: line 9: test1,test2
CONSOLE MESSAGE: line 10: [object NodeList]
CONSOLE MESSAGE: line 11: [object XPathResult]
Tests that console logging dumps proper messages.



console-dir.html:9Array[2]
    0: "test1"
    1: "test2"
    length: 2
    __proto__: Array[0]

console-dir.html:10NodeList[1]
    0: HTMLHtmlElement
    constructor: NodeListConstructor
    length: 1
    __proto__: NodeListPrototype

console-dir.html:11XPathResult
    booleanValue: XPathException
    constructor: XPathResultConstructor
    invalidIteratorState: false
    numberValue: XPathException
    resultType: 4
    singleNodeValue: XPathException
    snapshotLength: XPathException
    stringValue: XPathException
    __proto__: XPathResultPrototype
